CalcDropTime improved significantly
calcdroptime() was previously inaccurate because it did not account for the ball and paddle radius. With these factors now accounted for, the ball hits at the correct time down to approximately 0.01 sec. This should be good enough for our game, but a function may be added to ensure the ball snaps to the paddle at the correct time.